[syncherator] Make synching git repos with commits off-master work with shallow_since.
If a commit isn't reachable from master, then cloning/fetching with
shallow_since won't get the commit.

In that case, we fall back to fetching the commit itself; we rely on the
commit being advertised. (I think this is the case if the commit is
associated with a tag or branch head, but I'm doubtful about it
otherwise. Bazel doesn't let you specify branch+commit together though,
and it seems to do this sort of fallback.)